Factor out common interfaces
